migrate notifier to jsonschema
