+++ src/server/models/notifier/polymorphs/aws_sns_notifier_schema.json
@@ -1,0 +1,35 @@
+{
+  "title": "Notifier",
+  "type": "object",
+  "required": [
+    "name"
+  ],
+  "properties": {
+    "type": {
+      "type": "string",
+      "title": "watcher type. not editable",
+      "default": "CloudwatchAlarmWatcher"
+    },
+    "name": {
+      "type": "string",
+      "title": "name",
+      "description": "name of the component"
+    },
+    "awsAccessKeyId": {
+      "type": "string",
+      "title": "AWS AccessKeyId"
+    },
+    "awsSecretAccessKey": {
+      "type": "string",
+      "title": "AWS AccessKeySecret"
+    },
+    "awsRegion": {
+      "type": "string",
+      "title": "AWS Region"
+    },
+    "targetArn": {
+      "type": "string",
+      "title": "AWS SNS Target ARN"
+    }
+  }
+}
